Editted gitignore, added visual studio project
