+++ ObjectCopy/ObjectCopy.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2013
+VisualStudioVersion = 12.0.31101.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ObjectCopy", "ObjectCopy\ObjectCopy.csproj", "{5DE0B23C-E49A-4DA1-8225-4536BDEF8EC6}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{5DE0B23C-E49A-4DA1-8225-4536BDEF8EC6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5DE0B23C-E49A-4DA1-8225-4536BDEF8EC6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5DE0B23C-E49A-4DA1-8225-4536BDEF8EC6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5DE0B23C-E49A-4DA1-8225-4536BDEF8EC6}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
