+++ ObjectCopy/ObjectCopy/Program.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ObjectCopy
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+        }
+    }
+}
